% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structuring.R
\name{greet}
\alias{greet}
\title{Greet a person}
\usage{
greet(name, excited = FALSE)
}
\arguments{
\item{name}{A character string with the person's name.}

\item{excited}{Logical. Add exclamation marks? Default FALSE.}
}
\value{
A character string with the greeting.
}
\description{
Greet a person
}
\examples{
greet("R User", excited = TRUE)
}
